refactoring ItemKeyedBoundResource
